disable media for browsers that cannot support it

+++ src/app/messenger/web-sql.service.ts
@@ -1,18 +1,31 @@
 import { Injectable } from '@angular/core';
 
+import { AlertService } from '../alert.service';
+
 @Injectable({
   providedIn: 'root'
 })
 export class WebSqlService {
   private db: any;
   loaded = false;
+  supported = false;
 
-  constructor() { }
+  constructor(
+    private alertService: AlertService
+  ) { }
 
   init() {
     if ( !this.loaded ) {
       this.db = ( window as any ).openDatabase( 'lalilulelo', '1.0', 'Database For Managing Encrypted Client Side Data', 30 * 1024 * 1024 );
       this.loaded = true;
+    }
+
+    if ( !this.db ) {
+      console.log( 'feff' );
+      this.browserAlert();
+      return;
+    } else {
+      this.supported = true;
     }
 
     this.db.transaction( ( tx ) => {
@@ -22,6 +35,10 @@ export class WebSqlService {
 
   getMedia( messageId: string ) {
     return new Promise<any>( ( resolve, reject ) => {
+      if ( !this.db ) {
+        return resolve( null );
+      }
+
       this.db.transaction( ( tx ) => {
         tx.executeSql( `SELECT medias.media FROM medias WHERE messageId="${ messageId }" LIMIT 1`, [], ( newTx, results ) => {
           if ( !results.rows.length ) {
@@ -37,6 +54,10 @@ export class WebSqlService {
   }
 
   addMedia( messageId: string, media: string, expiration: number ) {
+    if ( !this.db ) {
+      return;
+    }
+
     this.db.transaction( ( tx ) => {
       tx.executeSql(`INSERT INTO medias VALUES('${ messageId }','${ media }',${ expiration })`, [], ( newTx ) => {
       }, ( newTx, err ) => {
@@ -46,6 +67,11 @@ export class WebSqlService {
   }
 
   removeMedia( messageId: string ) {
+    if ( !this.db ) {
+      this.browserAlert();
+      return;
+    }
+
     this.db.transaction( ( tx ) => {
       tx.executeSql(`DELETE FROM medias WHERE messageId="${ messageId }"`);
     } );
@@ -53,6 +79,10 @@ export class WebSqlService {
 
   getExpirations() {
     return new Promise<any>( ( resolve, reject ) => {
+      if ( !this.db ) {
+        return resolve( null );
+      }
+
       this.db.transaction( ( tx ) => {
         tx.executeSql( 'SELECT medias.messageId, medias.expiration FROM medias', [], ( newTx, results ) => {
           return resolve( results.rows );
@@ -63,41 +93,7 @@ export class WebSqlService {
     } );
   }
 
-private ab2str(buf) {
-  return String.fromCharCode.apply(null, new Uint16Array(buf));
-}
-
-private str2ab(str) {
-  const buf = new ArrayBuffer(str.length * 2); // 2 bytes for each char
-  const bufView = new Uint16Array(buf);
-  for (let i = 0, strLen = str.length; i < strLen; i++) {
-    bufView[i] = str.charCodeAt(i);
+  private browserAlert() {
+    this.alertService.alertSubject.next({ mode: 'danger', message: 'Your Browser Does Not Support WebSQL, Media Will Not Work.' });
   }
-  return buf;
-}
-
-private  mysql_real_escape_string (str) {
-    return str.replace(/[\0\x08\x09\x1a\n\r"'\\\%]/g, function (char) {
-        switch (char) {
-            case "\0":
-                return "\\0";
-            case "\x08":
-                return "\\b";
-            case "\x09":
-                return "\\t";
-            case "\x1a":
-                return "\\z";
-            case "\n":
-                return "\\n";
-            case "\r":
-                return "\\r";
-            case "\"":
-            case "'":
-            case "\\":
-            case "%":
-                return "\\"+char; // prepends a backslash to backslash, percent,
-                                  // and double/single quotes
-        }
-    });
-}
 }
